% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_bounds}
\alias{calculate_bounds}
\title{Calculate Confidence Interval Bounds}
\usage{
calculate_bounds(period, bootstraps, index, sum, sum_square)
}
\arguments{
\item{period}{A vector indicating the period associated with the data.}

\item{bootstraps}{An integer representing the number of bootstrap samples.}

\item{index}{The estimated index value.}

\item{sum}{The sum of bootstrap estimates.}

\item{sum_square}{The sum of squared bootstrap estimates.}
}
\value{
A data frame with the period, variance, lower bound, and upper bound of the confidence interval.
}
\description{
This internal function calculates the confidence interval bounds based on bootstrap estimates.
}
\keyword{internal}
