% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_geometric_average}
\alias{calculate_geometric_average}
\title{Calculate the geometric average of a series of values}
\usage{
calculate_geometric_average(values)
}
\arguments{
\item{values}{= series with numeric values}
}
\value{
geometric average
}
\description{
The equation for the calculation is:: exp(mean(log(reeks_values)))
}
\examples{
a <- c(1,2,3,4)
calculate_geometric_average(a)
# [1] 2.213364
}
\author{
Farley Ishaak (FIHK)
}
