% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_method_index.R
\name{calculate_method_index}
\alias{calculate_method_index}
\title{Calculate index based on specified method (Fisher, Laspeyres, Paasche, HMTS)}
\usage{
calculate_method_index(
  method,
  dataset,
  period_variable,
  dependent_variable,
  continuous_variables,
  categorical_variables,
  log_dependent = TRUE,
  reference_period = NULL,
  number_of_observations = TRUE,
  periods_in_year = 4,
  production_since = NULL,
  number_preliminary_periods = 3,
  resting_points = FALSE,
  index = TRUE,
  imputation = FALSE
)
}
\arguments{
\item{method}{One of: "fisher", "laspeyres", "paasche", "hmts"}

\item{dataset}{Data frame with input data}

\item{period_variable}{Name of the variable indicating time periods}

\item{dependent_variable}{Usually the price}

\item{continuous_variables}{Vector with numeric quality-determining variables}

\item{categorical_variables}{Vector with categorical variables (also dummies)}

\item{log_dependent}{Should the dependent variable be log-transformed? Default = TRUE}

\item{reference_period}{Period or group of periods that will be set to 100}

\item{number_of_observations}{Show number of observations? Default = TRUE}

\item{periods_in_year}{(HMTS only) Number of periods per year (e.g. 12 for months)}

\item{production_since}{(HMTS only) Start period for production simulation}

\item{number_preliminary_periods}{(HMTS only) Number of preliminary periods}

\item{resting_points}{(HMTS only) Return detailed outputs? Default = FALSE}

\item{index}{(Laspeyres/Paasche only) Include index column? Default = TRUE}

\item{imputation}{(Laspeyres/Paasche only) Include imputation values? Default = FALSE}
}
\value{
A data.frame (or list for HMTS with resting_points = TRUE)
}
\description{
Central hub function to calculate index figures using different methods.
}
\examples{
Tbl_Laspeyres <- calculate_method_index(
  method = "laspeyres",
  dataset = data_constraxion,
  period_variable = "period",
  dependent_variable = "price",
  continuous_variables = "floor_area",
  categorical_variables = "neighbourhood_code",
  log_dependent = TRUE,
  reference_period = 2015,
  number_of_observations = TRUE,
  imputation = TRUE
)
}
\author{
Vivek Gajadhar (VGAR)
}
