% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calculate_price_index}
\alias{calculate_price_index}
\title{Calculate Price Index (Deprecated)}
\usage{
calculate_price_index(...)
}
\arguments{
\item{...}{Arguments passed to \code{calculate_hedonic_index}.}
}
\description{
This function has been renamed. Please use \code{\link{calculate_hedonic_index}} instead.
}
