% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_timedummy.R
\name{calculate_time_dummy_index}
\alias{calculate_time_dummy_index}
\title{Calculate Time Dummy Index}
\usage{
calculate_time_dummy_index(
  dataset,
  period_variable,
  dependent_variable,
  continuous_variables,
  categorical_variables,
  reference_period = NULL,
  diagnostics = FALSE
)
}
\arguments{
\item{dataset}{data frame with input data}

\item{period_variable}{name of the time variable (string)}

\item{dependent_variable}{name of the dependent variable (usually price, assumed unlogged)}

\item{continuous_variables}{vector of numeric quality-determining variables}

\item{categorical_variables}{vector of categorical variables}

\item{reference_period}{period to be normalized to index = 100 (e.g., "2015")}

\item{diagnostics}{logical, whether to return number of observations per period (default = FALSE)}
}
\value{
data frame with period, Index, and optionally diagnostics
}
\description{
Estimates a price index using a single regression with time dummy variables.
}
\author{
Vivek Gajadhar
}
\keyword{internal}
