% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helpers.R
\name{custom_update_function}
\alias{custom_update_function}
\title{Default update function}
\usage{
custom_update_function(params, model)
}
\arguments{
\item{params}{startvalues}

\item{model}{state space modelnumber}
}
\value{
Newmodel
}
\description{
This function is used in the function: calculate_trend_line_KFAS()
}
\author{
Vivek Gajadhar
}
\keyword{internal}
