% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helpers.R
\name{defaultupdatefn}
\alias{defaultupdatefn}
\title{Default update function}
\usage{
defaultupdatefn(pars, model)
}
\arguments{
\item{pars}{= startvalues}

\item{model}{= state space modelnumber}
}
\value{
newmodel
}
\description{
This function is used in the function: calculate_trend_line_KFAS()
}
\author{
Pim Ouwehand
}
\keyword{internal}
