% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helpers.R
\name{determine_initial_parameters}
\alias{determine_initial_parameters}
\title{Determine_initial_parameters}
\usage{
determine_initial_parameters(model, initial_values, FUN = defaultupdatefn)
}
\arguments{
\item{model}{= modelvalues as output of the function select_state_space_model()}

\item{initial_values}{= $initial.values as output of the model}

\item{FUN}{= function called: defaultupdatefn}
}
\value{
New initial startvalues
}
\description{
Determine startvalues within state space models
This function is used in the function: calculate_trend_line_KFAS()
}
\examples{
startvalues_old <- set_startvalues(log(0.1), log(0.1), log(0.1),log(0.1),log(0.1))
modelsel <- select_state_space_model(selection = modelnumber
                                            , series = Origineel_TS
                                            , regressors = regressor
                                            , initial.values.all = startvalues_old)
model_SS <- model <- modelsel$model
startvalues_temp <- modelsel$initial.values
startvalues_temp <- determine_initial_parameters(model, initial.values = startvalues_temp)
startvalues_new <- startvalues_temp$initial.values2
}
\author{
Pim Ouwehand, Farley Ishaak
}
\keyword{internal}
