% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimate.TS.parameters}
\alias{estimate.TS.parameters}
\title{Estimate time series parameters}
\usage{
estimate.TS.parameters(model, initial.values)
}
\arguments{
\item{model}{= model values as output of the function select_state_space_model()}

\item{initial.values}{= $initial.values as output of the model}
}
\value{
parameter for the time series model
}
\description{
Estimate parameters to estimate trend lines
This function is used in the function: calculate_trend_line_KFAS()#'
}
\author{
Pim Ouwehand, Farley Ishaak
}
\keyword{internal}
