% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_helpers.R
\name{fit_hedonic_model}
\alias{fit_hedonic_model}
\title{Fit a Hedonic Linear Model}
\usage{
fit_hedonic_model(dataset, dependent_variable, independent_variables)
}
\arguments{
\item{dataset}{A data.frame containing the variables.}

\item{dependent_variable}{A string specifying the name of the dependent variable.}

\item{independent_variables}{A character vector of all independent variables.}
}
\value{
A fitted \code{lm} object.
}
\description{
A centralized helper to construct the formula and fit the linear model
for hedonic index calculations.
}
\author{
Vivek Gajadhar
}
\keyword{internal}
