% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_helpers.R
\name{predict_hedonic}
\alias{predict_hedonic}
\title{Predict from a Hedonic Model}
\usage{
predict_hedonic(model, newdata)
}
\arguments{
\item{model}{A fitted model object.}

\item{newdata}{A data.frame to predict on.}
}
\value{
A numeric vector of predictions.
}
\description{
Centralized prediction helper for hedonic models.
}
\author{
Vivek Gajadhar
}
\keyword{internal}
