% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helpers.R
\name{set_startvalues}
\alias{set_startvalues}
\title{Set starting values for hyperparameters in state space models}
\usage{
set_startvalues(a, b, c, d, e)
}
\value{
starting values for hyperparameters
}
\description{
Set starting values for hyperparameters in state space models
}
\keyword{internal}
