% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_progress_loop}
\alias{show_progress_loop}
\title{Title}
\usage{
show_progress_loop(single_iteration, total_iterations)
}
\arguments{
\item{single_iteration}{= a single iteration (usually 1 letter: i or p)}

\item{total_iterations}{= the total number of iterations in the loop}
}
\value{
this returns a progress text
}
\description{
Title
}
\keyword{internal}
