% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helpers.R
\name{smooth_ts}
\alias{smooth_ts}
\title{Run forward and backward pass of time series estimation}
\usage{
smooth_ts(fittedmodel)
}
\arguments{
\item{fittedmodel}{= model values as output of the function estimate.TS.parameters()}
}
\value{
sub-list $signalsubconf[,1] provides the estimated trend line
}
\description{
Calculate a trend line based on a provided model.
This function is used in the function: calculate_trend_line_KFAS()
}
\author{
Pim Ouwehand, Farley Ishaak
}
\keyword{internal}
