% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_hedonic_imputation}
\alias{calculate_hedonic_imputation}
\title{Title Calculation of a hedonic imputation}
\usage{
calculate_hedonic_imputation(
  dataset_temp,
  Period_temp,
  dependent_variable_temp,
  independent_variables_temp,
  log_dependent_temp,
  number_of_observations_temp,
  period_list_temp
)
}
\arguments{
\item{dataset_temp}{= table with data (hoeft no selectie te zijn van benodigde variables)}

\item{Period_temp}{= 'Period'}

\item{dependent_variable_temp}{= usually the sale price}

\item{independent_variables_temp}{= vector with quality determining variables}

\item{log_dependent_temp}{= should the dependent variable be transformed to its logarithm? default = TRUE}

\item{period_list_temp}{= list with all available periods}
}
\value{
table with imputation averages per period
}
\description{
Title Calculation of a hedonic imputation
}
\author{
Farley Ishaak
}
\keyword{internal}
