% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasp.R
\name{calculate_laspeyres}
\alias{calculate_laspeyres}
\title{Calculate direct index according to the Laspeyres hedonic double imputation method}
\usage{
calculate_laspeyres(
  dataset,
  period_variable,
  dependent_variable,
  continious_variables,
  categorical_variables,
  log_dependent = FALSE,
  reference_period = NULL,
  index = TRUE,
  number_of_observations = FALSE,
  imputation = FALSE
)
}
\arguments{
\item{dataset}{table with data (does not need to be a selection of relevant variables)}

\item{period_variable}{variable in the table with periods}

\item{dependent_variable}{usually the sale price}

\item{continious_variables}{vector with quality determining numeric variables (no dummies)}

\item{categorical_variables}{vector with quality determining categorical variables (also dummies)}

\item{log_dependent}{should the dependent variable be transformed to its logarithm? default = TRUE}

\item{reference_period}{period or group of periods that will be set to 100 (numeric/string)}

\item{index}{caprice index}

\item{number_of_observations}{number of observations per period (default = TRUE)}

\item{imputation}{display the inderlying average imputation values? (default = FALSE)}
}
\value{
table with index, imputation averages, number of observations and confidence intervals per period
}
\description{
By the parameters 'dependent_variable', 'continue_variable' and 'categorical_variables' as regression model is compiled.
With the model, a direct series of index figures is estimated by use of hedonic regression.
}
\details{
N.B.: the independent variables must be entered transformed (and ready) in the parameters.
Hence, not: log(floor_area), but transform the variable in advance and then provide log_floor_area.
This does not count for the dependent variable. This should be entered untransformed/
The parameter log_dependent can be used to transform this variable.

Within the data, it is not necessary to filter the data on relevant variables or complete records.
This is taken care of in the function.
}
\examples{

Tbl_Laspeyres <- calculate_laspeyres(dataset = data_constraxion
                                , period_variable = c("period")
                                , dependent_variable = c('price')
                                , continious_variables = c('floor_area')
                                , categorical_variables = c('neighbourhood_code')
                                , log_dependent = TRUE
                                , reference_period = 2015
                                , number_of_observations = TRUE
                                , imputation = TRUE)
}
\author{
Farley Ishaak (FIHK)
}
