% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calculate_price_index}
\alias{calculate_price_index}
\title{Calculate Price Index (Defunct)}
\usage{
calculate_price_index(...)
}
\arguments{
\item{...}{Arguments are ignored as the function is defunct.}
}
\description{
This function has been removed and replaced. Please use \code{\link{calculate_hedonic_index}} instead.
}
