% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_price_index.R
\name{calculate_price_index}
\alias{calculate_price_index}
\title{Calculate index based on specified method (Fisher, Laspeyres, Paasche, HMTS, Time Dummy, Rolling Time Dummy)}
\usage{
calculate_price_index(
  method,
  dataset,
  period_variable,
  dependent_variable,
  continuous_variables,
  categorical_variables,
  reference_period = NULL,
  number_of_observations = TRUE,
  periods_in_year = 4,
  production_since = NULL,
  number_preliminary_periods = 3,
  resting_points = FALSE,
  index = TRUE,
  imputation = FALSE,
  window_length = NULL
)
}
\arguments{
\item{method}{One of: "fisher", "laspeyres", "paasche", "hmts", "timedummy", "rolling_timedummy"}

\item{dataset}{Data frame with input data}

\item{period_variable}{Name of the variable indicating time periods}

\item{dependent_variable}{Usually the price}

\item{continuous_variables}{Vector with numeric quality-determining variables}

\item{categorical_variables}{Vector with categorical variables (also dummies)}

\item{reference_period}{Period or group of periods that will be set to 100}

\item{number_of_observations}{Logical, whether to show number of observations (default = TRUE)}

\item{periods_in_year}{(HMTS only) Number of periods per year (e.g. 12 for months)}

\item{production_since}{(HMTS only) Start period for production simulation}

\item{number_preliminary_periods}{(HMTS only) Number of preliminary periods}

\item{resting_points}{(HMTS only) Whether to return detailed outputs (default = FALSE)}

\item{index}{(Laspeyres/Paasche only) Include index column? Default = TRUE}

\item{imputation}{(Laspeyres/Paasche only) Include imputation values? Default = FALSE}

\item{window_length}{(Rolling Time Dummy only) Window size in number of periods}
}
\value{
A data.frame (or list for HMTS with resting_points = TRUE)
}
\description{
Central hub function to calculate index figures using different methods.
}
\examples{
# Laspeyres index
Tbl_Laspeyres <- calculate_price_index(
  method = "laspeyres",
  dataset = data_constraxion,
  period_variable = "period",
  dependent_variable = "price",
  continuous_variables = "floor_area",
  categorical_variables = "neighbourhood_code",
  reference_period = "2015",
  number_of_observations = FALSE,
  imputation = FALSE
)

# Paasche index
Tbl_Paasche <- calculate_price_index(
  method = "paasche",
  dataset = data_constraxion,
  period_variable = "period",
  dependent_variable = "price",
  continuous_variables = "floor_area",
  categorical_variables = "neighbourhood_code",
  reference_period = "2015",
  number_of_observations = FALSE,
  imputation = FALSE
)

# Fisher index (geometric mean of Laspeyres and Paasche)
Tbl_Fisher <- calculate_price_index(
  method = "fisher",
  dataset = data_constraxion,
  period_variable = "period",
  dependent_variable = "price",
  continuous_variables = "floor_area",
  categorical_variables = "neighbourhood_code",
  reference_period = "2015",
  number_of_observations = FALSE
)

# Time Dummy index
Tbl_TD <- calculate_price_index(
  method = "timedummy",
  dataset = data_constraxion,
  period_variable = "period",
  dependent_variable = "price",
  continuous_variables = "floor_area",
  categorical_variables = "neighbourhood_code",
  reference_period = "2015",
  number_of_observations = FALSE
)

# Rolling Time Dummy index
Tbl_TD_Rolling <- calculate_price_index(
  method = "rolling_timedummy",
  dataset = data_constraxion,
  period_variable = "period",
  dependent_variable = "price",
  continuous_variables = "floor_area",
  categorical_variables = "neighbourhood_code",
  reference_period = "2015",
  window_length = 5,
  number_of_observations = FALSE
)
}
