% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_trend_line_KFAS}
\alias{calculate_trend_line_KFAS}
\title{Calculate the trend line for a provided time series of numeric values}
\usage{
calculate_trend_line_KFAS(
  original_series,
  periodicity = 4,
  modelnumber = 5,
  logarithmic = TRUE,
  regressor = NULL,
  startvalues = NULL,
  resting_points = FALSE
)
}
\arguments{
\item{original_series}{= time series with values in chrolological order}

\item{periodicity}{= if month, then  12. If quarter, then 4, etc. (defaul = 4)}

\item{modelnumber}{= the modelnumber for the trend calcualtion (default = 5 -> smooth trend) No trend = NULL}

\item{logarithmic}{= logarithm of the dependent variable? (default = TRUE)}

\item{regressor}{= optional external variable that influences the series (default = NULL)}

\item{startvalues}{=  startvalues for 5 hyperparameters: meas, level, slope, seas, scaling. default = NULL: optimal estimation of these values.}

\item{resting_points}{= Should analyses values be returned? (default = FALSE)}
}
\value{
Trend line
}
\description{
Calculate the trend line with the state space method for a provided time series (chronological order is assumed).
The series are calculated with the package KFAS.
}
\author{
Pim Ouwehand, Farley Ishaak
}
