% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helpers.R
\name{calculate_trend_line_kfas}
\alias{calculate_trend_line_kfas}
\title{Calculate the trend line for a provided time series of numeric values}
\usage{
calculate_trend_line_kfas(original_series, periodicity, diagnostics)
}
\arguments{
\item{original_series}{time series with values in chrolological order}

\item{periodicity}{if month, then  12. If quarter, then 4, etc. (defaul = 4)}

\item{diagnostics}{should analyses values be returned? (default = FALSE)}
}
\value{
Trend line
}
\description{
Calculate the trend line with the state space method for a provided time series (chronological order is assumed).
The series are calculated with the package KFAS.
}
\author{
Pim Ouwehand, Farley Ishaak
}
\keyword{internal}
