% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_helpers.R
\name{determine_initial_parameters}
\alias{determine_initial_parameters}
\title{Determine_initial_parameters}
\usage{
determine_initial_parameters(
  model,
  initial_values,
  FUN = custom_update_function
)
}
\arguments{
\item{model}{modelvalues as output of the function select_state_space_model()}

\item{initial_values}{$initial.values as output of the model}

\item{FUN}{function called: custom_update_function}
}
\value{
New initial startvalues
}
\description{
Determine startvalues within state space models
This function is used in the function: calculate_trend_line_KFAS()
}
\author{
Pim Ouwehand, Farley Ishaak
}
\keyword{internal}
