% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_price_index.R
\name{plot_price_index}
\alias{plot_price_index}
\title{Plot index output from calculate_price_index}
\usage{
plot_price_index(index_output, title = NULL)
}
\arguments{
\item{index_output}{a data.frame (single method) or named list of data.frames (multi-method) from \code{calculate_price_index()}.}

\item{title}{optional custom plot title.}
}
\value{
a ggplot object
}
\description{
Plots one or more price indices over time. If multiple methods are provided, a legend is included.
X-axis shows only the first period of each year for clarity.
}
\author{
Vivek Gajadhar
}
