% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_state_space_model}
\alias{select_state_space_model}
\title{Select the state space model type}
\usage{
select_state_space_model(
  selection = 5,
  series,
  regressors = NULL,
  initial.values.all
)
}
\arguments{
\item{selection}{= the model number for the trend calculation (Default = 5).}

\item{series}{= time series with values in chronological order}

\item{regressors}{= optional external variable which influences the series (defaul = NULL)}

\item{initial.values.all}{= =  startvalues for 5 hyperparameters: meas, level, slope, seas, scaling}
}
\value{
modelvalues (level, slope) of the chosen state space model and the provided time series
}
\description{
This function is used in the function: calculate_trend_line_KFAS()
}
\author{
Pim Ouwehand
}
\keyword{internal}
