% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_startvalues}
\alias{set_startvalues}
\title{Title}
\usage{
set_startvalues(a, b, c, d, e)
}
\arguments{
\item{a}{mean}

\item{b}{level}

\item{c}{slope}

\item{d}{seas}

\item{e}{scaling}
}
\value{
starting values for hyperparameters
}
\description{
Title
}
\keyword{internal}
